monitor_service/requirements.txt
fastapi
uvicorn[standard]
google-cloud-spanner
httpx